// webui/src/gen_client/models/LayerCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type LayerCreate = {
  name?: string | null
  x?: number | null
  y?: number | null
  width?: number | null
  height?: number | null
  rotation?: number | null
  uuid?: string | null
  image_data?: string | null
}